feat: Move test schema and backend into main package
